continue the large refactoring to replace _hub by smth that forbids pygame
