+++ src/pyved_engine/__init__.py
@@ -11,6 +11,7 @@
 +----------------------------------------------------+
 """
 from . import _hub
+from .EngineRouter import EngineRouter
 from .Singleton import Singleton
 from ._classes import *
 from .compo import gfx
@@ -18,15 +19,14 @@ from .compo import vscreen  # deprecated
 from .compo.GameTpl import GameTpl  # legacy cls
 from .foundation import legacy_evs  # we just copy the event system of pygame
 from .foundation.events import Emitter, EvListener, EngineEvTypes
-from .utils import vars as defs
-from .EngineRouter import EngineRouter
+from . import pe_vars as defs
 
 
 # TODO remove this when we can
 # ive kept it for retro-compatibility with projects that target pyv v23.6a1
 # such as demos/ecs_naif or the very early stage pyved ships-with-GUI editor
-from .utils._ecs_pattern import entity, component, System, SystemManager, EntityManager
-from . import evsys0
+# from .utils._ecs_pattern import entity, component, System, SystemManager, EntityManager
+# from . import evsys0
 
 
 _stored_kbackend = None
@@ -37,16 +37,3 @@ def set_webbackend_type(xval):
     global _backend_name
     vars.weblib_sig = _backend_name = xval
     vars.backend_name = 'web'
-
-
-# the basic API used to be expanded via our special "hub" component
-# def __getattr__(attr_name):
-#     if attr_name in ('ver', 'vernum'):
-#         return get_version()
-#     elif attr_name == 'Sprite':
-#         return dep_linking.pygame.sprite.Sprite
-#     elif attr_name == 'SpriteGroup':
-#         return dep_linking.pygame.sprite.Group
-#     elif attr_name == 'sprite_collision':
-#         return dep_linking.pygame.sprite.spritecollide
-#     return getattr(_hub, attr_name)
