+++ src/pyved_engine/foundation/pbackends.py
@@ -1,7 +1,7 @@
 from .defs import EngineEvTypes, to_camelcase
 from .interfaces import BaseKenBackend
 from .. import _hub
-from ..utils import vars
+from .. import pe_vars
 from .. import dep_linking
