--- src/pyved_engine/gamedev_api/local_context.py
@@ -1,12 +1,0 @@
-
-
-def engine_init(mode):
-    print('engine_init local context!')
-
-
-def machin():
-    print('machin local context!')
-
-
-def flip():
-    print('flip local context!')
