+++ src/pyved_engine/looparts/polarbear/frects.py
@@ -1,6 +1,6 @@
-from ... import _hub
-from ... import vars
+from ... import pe_vars as _vars
 
+pyv = _vars.engine
 
 ANCHOR_UPPERLEFT = (0, 0)
 ANCHOR_UPPERRIGHT = (2, 0)
@@ -10,8 +10,6 @@ ANCHOR_LOWERRIGHT = (2, 2)
 ANCHOR_TOP = (1, 0)
 ANCHOR_LEFT = (0, 1)
 ANCHOR_RIGHT = (2, 1)
-
-pygame = _hub.pygame
 
 
 class Frect(object):
@@ -31,7 +29,7 @@ class Frect(object):
             x0 = prect.left + (prect.w // 2) * self.anchor[0]
             y0 = prect.top + (prect.h // 2) * self.anchor[1]
         else:
-            sw, sh = vars.screen.get_size()
+            sw, sh = _vars.screen.get_size()
             x0 = (sw // 2) * self.anchor[0]
             y0 = (sh // 2) * self.anchor[1]
-        return pygame.Rect(self.dx + x0, self.dy + y0, self.w, self.h)
+        return pyv.new_rect_obj(self.dx + x0, self.dy + y0, self.w, self.h)
